requirements.txt
asgiref==3.2.3
cffi==1.13.2
cryptography==2.8
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.1.1
Pillow==7.0.0
pycparser==2.19
PyMySQL==0.9.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
